+++ functional/LazyLoadImage.svelte
@@ -1,10 +1,11 @@
 <script>
-	export let src, alt;
+	export let src = '';
+	export let alt = '';
 	export let overlay = false;
 	export let absolute = false;
 
-	import Observe from './Observe.svelte';
-	import Image from './Image.svelte';
+	import Observe from '../essentials/Observe.svelte';
+	import Image from '../essentials/Image.svelte';
 </script>
 
 <Observe once let:sighted>