+++ functional/index.d.ts
@@ -1,0 +1,4 @@
+export { default as LazyLoadImage } from './LazyLoadImage.svelte';
+export { default as Pagination } from './Pagination.svelte';
+export { default as SearchBar } from './SearchBar.svelte';
+export { default as ThemeSwitcher } from './ThemeSwitcher.svelte';