+++ rollup.config.js
@@ -1,5 +1,6 @@
-import svelte from 'rollup-plugin-svelte';
 import resolve from '@rollup/plugin-node-resolve';
+import svelte from 'rollup-plugin-svelte';
+import autoPreprocess from 'svelte-preprocess';
 import pkg from './package.json';
 
 const name = pkg.name
@@ -8,10 +9,10 @@ const name = pkg.name
 	.replace(/-\w/g, (m) => m[1].toUpperCase());
 
 export default {
-	input: 'svelte/index.js',
+	input: 'index.js',
 	output: [
 		{ file: pkg.module, format: 'es' },
 		{ file: pkg.main, format: 'umd', name },
 	],
-	plugins: [svelte(), resolve()],
+	plugins: [svelte({ preprocess: autoPreprocess() }), resolve()],
 };