+++ styled/ButtonLink.svelte
@@ -5,10 +5,10 @@
 	export let download = null;
 	export let label = null;
 	export let invert = false;
-	import Link from './Link.svelte';
+	import Link from '../essentials/Link.svelte';
 </script>
 
-<span class="elements button-link" class:disabled>
+<span class="elements styled button-link" class:disabled>
 	<Link href={!disabled ? href : ''} inherit {newTab} {label} {download} {invert}>
 		<slot />
 	</Link>